There we go. Probably lots of bugs in this code.
Created an ExecutorService in the Client Creator. However, should there be a "Client State" class?

Incorporated the Client FXML, which is pretty much clear at this point, into the ClientCreator.

Delegates the Executor Service from... somewhere

Created the Queuer class
Created an ExecutorService in ClientMain

Optimized the SocketHandler with Queuer